+++ src/main/psk-ui/src/constants/menuListConstants.ts
@@ -1,0 +1,10 @@
+//MENU LIST ITEM NAMES
+export const CALENDAR_MENU_ITEM_NAME = 'My Calendar';
+export const USER_MENU_ITEM_NAME = 'My Profile';
+export const MY_TEAM_CALENDAR_MENU_ITEM_NAME = 'My Team Calendar';
+export const INFO_MENU_ITEM_NAME = 'Info';
+export const TEAMS_MENU_ITEM_NAME = 'Teams';
+export const TEAM_MEMBERS_MENU_ITEM_NAME = 'Team Members';
+export const TOPICS_MENU_ITEM_NAME = 'Topics';
+export const TOPIC_TREE_MENU_ITEM_NAME = 'Topic Tree';
+export const CREATE_NEW_TOPIC_MENU_ITEM_NAME = 'New Topic';
