+++ src/main/psk-ui/src/constants/routeKeyConstants.ts
@@ -1,0 +1,12 @@
+// KEYS
+export const KEY_TEAMS = 'teams';
+export const KEY_CALENDAR = 'calendar';
+export const KEY_MY_CALENDAR = 'my-calendar';
+export const KEY_MEMBERS = 'members';
+export const KEY_MY_EMPLOYEES= 'my-employees';
+export const KEY_TEAM = 'team';
+export const KEY_PROFILE = 'profile';
+export const KEY_INFO = 'info';
+export const KEY_TOPICS = 'topics';
+export const KEY_TOPIC_TREE = 'topic-tree';
+export const KEY_NEW_TOPIC = 'new-topic';
