+++ src/main/psk-ui/src/pages/private/profile-view/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Button, Card, Timeline, Typography} from 'antd';
+import { Card, Timeline, Typography } from 'antd';
 // this is for "My Profile"
 // But may(?) be used for viewing individual profiles of employees
 import './ProfileStyles.css';
