+++ src/main/psk-ui/src/pages/private/team-members-view/editable-table/EditableTableStyles.css
@@ -4,7 +4,3 @@
 .infoAlert {
     margin-bottom: 16px !important;
 }
-li {
-    font-size: 14px;
-    line-height: 22px;
-}
